requirements/test.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements/test.txt requirements/test.in
#
asgiref==3.8.1
    # via
    #   -r requirements/base.txt
    #   django
cffi==1.17.1
    # via pynacl
click==8.1.6
    # via
    #   -c requirements/constraints.txt
    #   code-annotations
    #   edx-django-utils
code-annotations==2.3.0
    # via -r requirements/test.in
coverage[toml]==7.8.0
    # via pytest-cov
ddt==1.7.2
    # via -r requirements/test.in
deepdiff==8.5.0
    # via -r requirements/test.in
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/base.txt
    #   django-config-models
    #   django-crum
    #   django-filter
    #   django-mysql
    #   django-waffle
    #   djangorestframework
    #   drf-yasg
    #   edx-api-doc-tools
    #   edx-django-utils
    #   jsonfield
django-config-models==2.9.0
    # via -r requirements/test.in
django-crum==0.7.9
    # via edx-django-utils
django-filter==25.1
    # via -r requirements/test.in
django-mysql==4.17.0
    # via -r requirements/test.in
django-simple-history==3.0.0
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/test.in
django-waffle==4.2.0
    # via edx-django-utils
djangorestframework==3.16.0
    # via
    #   -r requirements/test.in
    #   django-config-models
    #   drf-yasg
    #   edx-api-doc-tools
dnspython==2.7.0
    # via pymongo
drf-yasg==1.21.10
    # via edx-api-doc-tools
edx-api-doc-tools==2.1.0
    # via -r requirements/test.in
edx-django-utils==8.0.0
    # via django-config-models
edx-opaque-keys[django]==3.0.0
    # via -r requirements/test.in
inflection==0.5.1
    # via drf-yasg
iniconfig==2.1.0
    # via pytest
jinja2==3.1.6
    # via code-annotations
jsonfield==3.1.0
    # via -r requirements/test.in
markupsafe==3.0.2
    # via jinja2
numpy==2.2.6
    # via -r requirements/test.in
openedx-atlas==0.7.0
    # via -r requirements/base.txt
orderly-set==5.4.1
    # via deepdiff
packaging==25.0
    # via
    #   drf-yasg
    #   pytest
pbr==6.1.1
    # via stevedore
pluggy==1.6.0
    # via pytest
psutil==7.0.0
    # via edx-django-utils
pycparser==2.22
    # via cffi
pymongo==4.13.0
    # via edx-opaque-keys
pynacl==1.5.0
    # via edx-django-utils
pytest==8.3.5
    # via
    #   pytest-cov
    #   pytest-django
pytest-cov==6.1.1
    # via -r requirements/test.in
pytest-django==4.11.1
    # via -r requirements/test.in
python-dateutil==2.9.0.post0
    # via -r requirements/test.in
python-slugify==8.0.4
    # via code-annotations
pytz==2025.2
    # via drf-yasg
pyyaml==6.0.2
    # via
    #   code-annotations
    #   drf-yasg
six==1.17.0
    # via python-dateutil
sqlparse==0.5.3
    # via
    #   -r requirements/base.txt
    #   django
stevedore==5.4.1
    # via
    #   code-annotations
    #   edx-django-utils
    #   edx-opaque-keys
text-unidecode==1.3
    # via python-slugify
types-python-dateutil==2.9.0.20250516
    # via -r requirements/test.in
typing-extensions==4.13.2
    # via edx-opaque-keys
uritemplate==4.1.1
    # via drf-yasg

# The following packages are considered to be unsafe in a requirements file:
# setuptools
